Add files via upload
Its a customer segementation in which cust id,genre, age,annual income,spending score. We used feature selection and building the model by kmeans(clustering), it is a unsupervised technique, then this data is visualized by ELBOW Method, took the value of k = 5,which is on the point 
of the elbow of the graph, then build KMeans Model and then visualize all the clusters of the customer segmentation.